editor/libeditor/tests/test_bug587461.html

<!DOCTYPE html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=587461
-->
<title>Test for Bug 587461</title>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
<script src="/tests/SimpleTest/EventUtils.js"></script>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=587461">Mozilla Bug 587461</a>
<div contenteditable><b>foobar</b></div>
<script>
var div = document.querySelector("div");
getSelection().collapse(div.firstChild.firstChild, 3);
document.execCommand("inserthtml", false, "a");
is(div.innerHTML, "<b>fooabar</b>", "innerHTML");
</script>
